[HTTP-SERVLET EXAMPLE] :: Welcome to a username submitted from a form
As by default 'index.jsp' will be loaded automatically, I create there a
form to submit username in a '/Welcome' url. This request will be recei-
ved by tomcat and it will try to find the url mapping on 'web.xml' (dep-
loyment descriptor). And I create a Servlet mapping for '/Welcome' which
is a 'practice.servlet.WelcomeServlet'.

As I do not mention 'method' in 'index.jsp' form the request will be tr-
eated as a 'get' request hence 'doGet' method will be called.